:hammer: REFACTOR (user): move user api to page
