+++ src/layouts/User/index.tsx
@@ -1,21 +1,31 @@
 import React, { ReactElement } from "react";
 import { Link } from "react-router-dom";
 
-import ProfileViewContainer from "containers/ProfileView";
-import ProjectListContainer from "containers/ProjectList";
-import TagListContainer from "containers/TagList";
-import { ProjectFilterModel } from "models/Project";
-import { ProfileQueryModel } from "models/User";
+import ProfileView from "components/ProfileView";
+import ProjectList from "components/ProjectList";
+import TagList from "components/TagList";
+import ProjectModel from "models/Project";
+import TagModel from "models/Tag";
+import { ProfileModel } from "models/User";
+
+interface UserLayoutProps {
+  profile: ProfileModel;
+  tagObjects: TagModel[];
+  tags: number;
+  projects: ProjectModel[];
+}
 
 const UserLayout = ({
-  username,
+  profile,
+  tagObjects,
   tags,
-}: ProfileQueryModel & ProjectFilterModel): ReactElement => (
+  projects,
+}: UserLayoutProps): ReactElement => (
   <>
     <Link to="/">Home</Link>
-    <ProfileViewContainer username={username} />
-    <TagListContainer username={username} tags={tags} />
-    <ProjectListContainer username={username} tags={tags} />
+    <ProfileView {...profile} />
+    <TagList tags={tagObjects} currentFilter={tags} />
+    <ProjectList projects={projects} />
   </>
 );
